+++ database/migrations/2024_09_27_140019_add_groupe_ventes_id_to_ventes_table.php
@@ -4,21 +4,25 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class AddGroupeVentesIdToVentesTable extends Migration
 {
     public function up()
-{
-    Schema::table('ventes', function (Blueprint $table) {
-        $table->unsignedBigInteger('groupe_ventes_id')->nullable();
-        $table->foreign('groupe_ventes_id')->references('id')->on('groupe_ventes')->onDelete('cascade');
-    });
-}
+    {
+        Schema::table('ventes', function (Blueprint $table) {
+            // Vérifiez si la colonne existe avant de l'ajouter
+            if (!Schema::hasColumn('ventes', 'groupe_ventes_id')) {
+                $table->unsignedBigInteger('groupe_ventes_id')->nullable();
+            }
+        });
+    }
 
-public function down()
-{
-    Schema::table('ventes', function (Blueprint $table) {
-        $table->dropForeign(['groupe_ventes_id']);
-        $table->dropColumn('groupe_ventes_id');
-    });
+    public function down()
+    {
+        Schema::table('ventes', function (Blueprint $table) {
+            // Vérifiez si la colonne existe avant de tenter de la supprimer
+            if (Schema::hasColumn('ventes', 'groupe_ventes_id')) {
+                $table->dropColumn('groupe_ventes_id');
+            }
+        });
+    }
 }
-};
